requirements.txt
functorch==1.13.1
matplotlib==3.6.0
numpy==1.23.4
Pillow==9.4.0
PyYAML==6.0
torch==1.13.1
torchmetrics[image]==0.11.1
torchvision==0.14.1
tqdm==4.64.1
